requirements.txt
django==1.10.2
django-configurations==2.0
psycopg2==2.6.2
Pillow==3.4.2
djangorestframework==3.5.1
django-resized==0.3.5
factory_boy==2.7.0
